requirements.txt
Django==4.2.16
djangorestframework==3.16.1
Pillow==10.4.0
django-debug-toolbar==4.4.6
reportlab==4.2.5

# PostgreSQL support (раскомментировать при необходимости)
# psycopg2-binary==2.9.9

